requirements.txt
langchain
langchain-community
sentence-transformers
langchain_pinecone
pinecone-client
python-dotenv
streamlit
pypdf
nltk
scikit-learn 
rouge-score